Refactored code for opposite()

+++ lab04/calculator/src/CalculatorTest.java
@@ -21,7 +21,7 @@ class CalculatorTest {
 		assertEquals(Calculator.opposite(154), -451);
 		assertEquals(Calculator.opposite(307), -703);
 		assertEquals(Calculator.opposite(10000), -1);
-		assertEquals(Calculator.opposite(00093f), -39);
+		assertEquals(Calculator.opposite((int)00093f), -39);
 	}
 
 }
